requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-allauth==0.33.0
django-bootstrap3==9.0.0
django-crispy-forms==1.6.1
django-extensions==2.1.2
django-mailchimp-v1.3==1.4.0
idna==2.6
oauthlib==2.0.4
pbr==3.1.1
Pillow==5.2.0
psycopg2==2.7.3.1
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
stevedore==1.26.0
typing==3.6.6
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.2
